# Create SetPagetitleInView linter for haml-lint and enable it