--- a/src/api/app/views/webui/apidocs/index.html.haml
+++ b/src/api/app/views/webui/apidocs/index.html.haml
@@ -1,3 +1,5 @@
+- @pagetitle = 'API Documentation'
+
 .card
   .card-body
     -# rubocop:disable Rails/OutputSafety