--- a/src/api/app/views/webui/cloud/azure/configurations/show.html.haml
+++ b/src/api/app/views/webui/cloud/azure/configurations/show.html.haml
@@ -1,7 +1,8 @@
+- @pagetitle = 'Microsoft Azure Configuration'
+
 .card.mb-3
   .card-body
-    %h3.mb-3
-      Microsoft Azure Configuration
+    %h3.mb-3= @pagetitle
 
     %p
       You need to provide an application id and an application key in order to allow OBS to upload images to your Azure account.