--- a/src/api/app/views/webui/cloud/configurations/index.html.haml
+++ b/src/api/app/views/webui/cloud/configurations/index.html.haml
@@ -1,7 +1,8 @@
+- @pagetitle = 'Cloud Upload Configuration'
+
 .card.mb-3
   .card-body
-    %h3.mb-3
-      Cloud Upload Configuration
+    %h3.mb-3= @pagetitle
     %p
       You need to configure the cloud service providers you want to upload to.
     %ul