--- a/src/api/app/views/webui/cloud/ec2/configurations/show.html.haml
+++ b/src/api/app/views/webui/cloud/ec2/configurations/show.html.haml
@@ -1,7 +1,8 @@
+- @pagetitle = 'Amazon EC2 Configuration'
+
 .card.mb-3
   .card-body
-    %h3.mb-3
-      Amazon EC2 Configuration
+    %h3.mb-3= @pagetitle
     %p
       Amazon EC2 allows you to easily run instances of your appliance on Amazon's webservers.