--- a/src/api/app/views/webui/kiwi/images/show.html.haml
+++ b/src/api/app/views/webui/kiwi/images/show.html.haml
@@ -2,6 +2,7 @@
   @layouttype = 'custom'
   @project = @image.package.try(:project)
   @package = @image.package
+  @pagetitle = 'Kiwi Image'
 
 .row
   .col-12