--- a/src/api/app/views/webui/staging/excluded_requests/index.html.haml
+++ b/src/api/app/views/webui/staging/excluded_requests/index.html.haml
@@ -1,9 +1,10 @@
+- @pagetitle = 'Excluded Requests'
 - request_exclusion_policy = Staging::RequestExclusionPolicy.new(User.possibly_nobody, @staging_workflow).create?
 .row
   .col-xl-12
     .card.mb-3
       .card-body
-        %h3 Excluded Requests
+        %h3= @pagetitle
         -# haml-lint:disable MultilinePipe
         %table.responsive.table.table-sm.table-bordered.table-hover.w-100#excluded-requests-datatable{ |
         data: excluded_requests_remote_url(@staging_workflow.project) } |