--- a/src/api/app/views/webui/timeout.html.haml
+++ b/src/api/app/views/webui/timeout.html.haml
@@ -1,4 +1,6 @@
-%h1 Timeout occurred
+- @pagetitle = 'Timeout occurred'
+
+%h1= @pagetitle
 %p
   We're sorry but the system is currently not able to handle your request
   in a reasonable time. Please try again later.