--- a/src/api/app/views/webui/users/notifications/index.html.haml
+++ b/src/api/app/views/webui/users/notifications/index.html.haml
@@ -1,3 +1,5 @@
+- @pagetitle = 'Notifications'
+
 = render partial: 'index_actions'
 
 .row