## Summary

We need to set the instance variable `@pagetitle` in views to have a page title when the view is rendered. This custom haml-lint linter is there to remind us if we forget. A PR like #11449 wouldn't need to be created since it would be caught by this custom linter.

`@pagetitle` is used in the [webui layout](https://github.com/openSUSE/open-build-service/blob/032494fd268578c4dfe6f7b77a7b0bd71bef61fe/src/api/app/views/layouts/webui/webui.html.haml#L10).

## Benchmarks

I ran some simple benchmarks on my computer in our Docker development environment by executing `time haml-lint` with the custom linter enabled and disabled.

When the custom linter `SetPagetitleInView` is:
- enabled: 29.2 seconds to 30 seconds
- disabled: 29.6 seconds to 30.9 seconds

I'm okay with this super minor time increase in the CI if this prevents us from forgetting to set `@pagetitle` in views.